rename `pcm_optical` to `cdsem_all` and add `cdsem_coupler`
