+++ gdsfactory/components/cdsem_all.py
@@ -1,0 +1,89 @@
+"""CD SEM structures."""
+from typing import Optional, Tuple
+
+from gdsfactory.cell import cell
+from gdsfactory.component import Component
+from gdsfactory.components.bend_circular import bend_circular
+from gdsfactory.components.cdsem_bend180 import cdsem_bend180
+from gdsfactory.components.cdsem_straight import cdsem_straight
+from gdsfactory.components.cdsem_straight_density import cdsem_straight_density
+from gdsfactory.components.straight import straight as straight_function
+from gdsfactory.components.text_rectangular import text_rectangular
+from gdsfactory.cross_section import strip
+from gdsfactory.types import ComponentFactory, CrossSectionFactory
+
+
+@cell
+def cdsem_all(
+    widths: Tuple[float, ...] = (0.4, 0.45, 0.5, 0.6, 0.8, 1.0),
+    dense_lines_width: Optional[float] = 0.3,
+    dense_lines_width_difference: float = 20e-3,
+    dense_lines_gap: float = 0.3,
+    dense_lines_labels: Tuple[str, ...] = ("DL", "DM", "DH"),
+    straight: ComponentFactory = straight_function,
+    bend90: Optional[ComponentFactory] = bend_circular,
+    cross_section: CrossSectionFactory = strip,
+    text: ComponentFactory = text_rectangular,
+) -> Component:
+    """column with all optical PCMs
+
+    Args:
+        widths: for straight lines
+
+    """
+    c = Component()
+    _c1 = cdsem_straight(
+        widths=widths,
+        cross_section=cross_section,
+    )
+
+    all_devices = [_c1]
+
+    if bend90:
+        all_devices += [
+            cdsem_bend180(
+                width=width,
+                straight=straight,
+                bend90=bend90,
+                cross_section=cross_section,
+                text=text,
+            )
+            for width in widths
+        ]
+
+    if dense_lines_width:
+        density_params = [
+            (
+                dense_lines_width - dense_lines_width_difference,
+                dense_lines_gap - dense_lines_width_difference,
+                dense_lines_labels[0],
+            ),
+            (dense_lines_width, dense_lines_gap, dense_lines_labels[1]),
+            (
+                dense_lines_width + dense_lines_width_difference,
+                dense_lines_gap + dense_lines_width_difference,
+                dense_lines_labels[2],
+            ),
+        ]
+
+        all_devices += [
+            cdsem_straight_density(
+                width=w,
+                trench_width=t,
+                label=lbl,
+                straight=straight,
+                cross_section=cross_section,
+                text=text,
+            )
+            for w, t, lbl in density_params
+        ]
+
+    [c.add_ref(d) for d in all_devices]
+    c.align(elements="all", alignment="xmin")
+    c.distribute(elements="all", direction="y", spacing=5, separation=True)
+    return c
+
+
+if __name__ == "__main__":
+    c = cdsem_all()
+    c.show()
