--- gdsfactory/components/pcm_optical.py
@@ -1,253 +1,0 @@
-"""CD SEM structures."""
-from functools import partial
-from typing import List, Optional, Tuple
-
-from gdsfactory.cell import cell
-from gdsfactory.component import Component
-from gdsfactory.components.bend_circular import bend_circular
-from gdsfactory.components.coupler_straight import coupler_straight
-from gdsfactory.components.straight import straight as straight_function
-from gdsfactory.components.text_rectangular import text_rectangular
-from gdsfactory.cross_section import strip
-from gdsfactory.grid import grid
-from gdsfactory.tech import LAYER
-from gdsfactory.types import ComponentFactory, CrossSectionFactory, Layer
-
-LINE_LENGTH = 420.0
-
-
-@cell
-def cdsem_straight(
-    widths: Tuple[float, ...] = (0.4, 0.45, 0.5, 0.6, 0.8, 1.0),
-    length: float = LINE_LENGTH,
-    cross_section: CrossSectionFactory = strip,
-) -> Component:
-    """Returns straight waveguide lines width sweep.
-
-    Args:
-        widths: for the sweep
-        length: for the line
-        cross_section:
-    """
-
-    lines = []
-    for width in widths:
-        cross_section = partial(cross_section, width=width)
-        lines.append(straight_function(length=length, cross_section=cross_section))
-
-    return grid(lines)
-
-
-@cell
-def cdsem_coupler(
-    width: float = 0.45,
-    length: float = LINE_LENGTH,
-    gaps: Tuple[float, ...] = (0.15, 0.2, 0.25),
-    cross_section: CrossSectionFactory = strip,
-) -> Component:
-    """Returns 2 coupled waveguides gap sweep.
-
-    Args:
-        width: for the waveguide
-        length: for the line
-        gaps: list of gaps
-        cross_section
-
-    """
-
-    cross_section = partial(cross_section, width=width)
-    couplers = [
-        coupler_straight(length=length, gap=gap, cross_section=cross_section)
-        for gap in gaps
-    ]
-    return grid(couplers)
-
-
-@cell
-def cdsem_straight_density(
-    width: float = 0.3,
-    trench_width: float = 0.3,
-    x: float = LINE_LENGTH,
-    y: float = 50.0,
-    margin: float = 2.0,
-    label: str = "",
-    straight: ComponentFactory = straight_function,
-    layer: Tuple[int, int] = LAYER.WG,
-    layers_cladding: Optional[Tuple[Layer, ...]] = None,
-    cross_section: CrossSectionFactory = strip,
-    text: ComponentFactory = text_rectangular,
-) -> Component:
-    """Horizontal grating etch lines
-
-    Args:
-        width: width
-        trench_width: trench_width
-    """
-    c = Component()
-    period = width + trench_width
-    n_o_lines = int((y - 2 * margin) / period)
-    length = x - 2 * margin
-
-    cross_section = partial(cross_section, width=width, layer=layer)
-    tooth = straight(length=length, cross_section=cross_section)
-
-    for i in range(n_o_lines):
-        tooth_ref = c.add_ref(tooth)
-        tooth_ref.movey((-n_o_lines / 2 + 0.5 + i) * period)
-        c.absorb(tooth_ref)
-
-    marker_label = text(
-        text=f"{label}",
-        layer=layer,
-    )
-    _marker_label = c.add_ref(marker_label)
-    _marker_label.move((length + 3, 10.0))
-    c.absorb(_marker_label)
-    return c
-
-
-@cell
-def cdsem_bend180(
-    width: float = 0.5,
-    radius: float = 10.0,
-    wg_length: float = LINE_LENGTH,
-    straight: ComponentFactory = straight_function,
-    bend90: ComponentFactory = bend_circular,
-    layer: Tuple[int, int] = LAYER.WG,
-    layers_cladding: List[Tuple[int, int]] = None,
-    cross_section: CrossSectionFactory = strip,
-    text: ComponentFactory = text_rectangular,
-) -> Component:
-    """
-
-    Args:
-        width: of the line
-        cladding_offset:
-        radius: bend radius
-        wg_length
-
-    """
-    c = Component()
-    r = radius
-
-    cross_section = partial(cross_section, width=width, layer=layer)
-    if wg_length is None:
-        wg_length = 2 * r
-
-    bend90 = bend90(cross_section=cross_section, radius=r)
-    wg = straight(
-        cross_section=cross_section,
-        length=wg_length,
-    )
-
-    # Add the U-turn on straight layer
-    b1 = c.add_ref(bend90)
-    b2 = c.add_ref(bend90)
-    b2.connect("o2", b1.ports["o1"])
-
-    wg1 = c.add_ref(wg)
-    wg1.connect("o1", b1.ports["o2"])
-
-    wg2 = c.add_ref(wg)
-    wg2.connect("o1", b2.ports["o1"])
-
-    label = c << text(text=str(int(width * 1e3)), layer=layer)
-    label.ymax = b2.ymin - 5
-    label.x = 0
-    b1.rotate(90)
-    b2.rotate(90)
-    wg1.rotate(90)
-    wg2.rotate(90)
-    label.rotate(90)
-    return c
-
-
-@cell
-def pcm_optical(
-    widths: Tuple[float, ...] = (0.4, 0.45, 0.5, 0.6, 0.8, 1.0),
-    dense_lines_width: Optional[float] = 0.3,
-    dense_lines_width_difference: float = 20e-3,
-    dense_lines_gap: float = 0.3,
-    dense_lines_labels: Tuple[str, ...] = ("DL", "DM", "DH"),
-    straight: ComponentFactory = straight_function,
-    bend90: Optional[ComponentFactory] = bend_circular,
-    layer: Tuple[int, int] = LAYER.WG,
-    layers_cladding: List[Tuple[int, int]] = None,
-    cross_section: CrossSectionFactory = strip,
-    text: ComponentFactory = text_rectangular,
-) -> Component:
-    """column with all optical PCMs
-
-    Args:
-        widths: for straight lines
-
-    """
-    c = Component()
-    _c1 = cdsem_straight(
-        straight=straight,
-        layer=layer,
-        layers_cladding=layers_cladding,
-        widths=widths,
-        cross_section=cross_section,
-        text=text,
-    )
-
-    all_devices = [_c1]
-
-    if bend90:
-        all_devices += [
-            cdsem_bend180(
-                width=width,
-                straight=straight,
-                bend90=bend90,
-                layer=layer,
-                layers_cladding=layers_cladding,
-                cross_section=cross_section,
-                text=text,
-            )
-            for width in widths
-        ]
-
-    if dense_lines_width:
-        density_params = [
-            (
-                dense_lines_width - dense_lines_width_difference,
-                dense_lines_gap - dense_lines_width_difference,
-                dense_lines_labels[0],
-            ),
-            (dense_lines_width, dense_lines_gap, dense_lines_labels[1]),
-            (
-                dense_lines_width + dense_lines_width_difference,
-                dense_lines_gap + dense_lines_width_difference,
-                dense_lines_labels[2],
-            ),
-        ]
-
-        all_devices += [
-            cdsem_straight_density(
-                width=w,
-                trench_width=t,
-                label=lbl,
-                straight=straight,
-                layer=layer,
-                layers_cladding=layers_cladding,
-                cross_section=cross_section,
-                text=text,
-            )
-            for w, t, lbl in density_params
-        ]
-
-    [c.add_ref(d) for d in all_devices]
-    c.align(elements="all", alignment="xmin")
-    c.distribute(elements="all", direction="y", spacing=5, separation=True)
-    return c
-
-
-if __name__ == "__main__":
-    c = cdsem_coupler()
-    # c = cdsem_straight()
-    # c = cdsem_straight()
-    # c = cdsem_bend180(width=2)
-    # c = cdsem_straight_density()
-    # c = pcm_optical(layer=(2, 0))
-    c.show()
